refactor: remove webpack-env types and consolidate auth interfaces
